# docs/swagger.yaml
basePath: /
definitions:
  github_com_abduromanov2020_tasks-api_internal_usecase.TaskView:
    properties:
      assignee_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: string
      status:
        type: string
      team_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  internal_handler.AssignTaskRequest:
    properties:
      assignee_id:
        type: string
    required:
    - assignee_id
    type: object
  internal_handler.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/internal_handler.UserResponse'
    type: object
  internal_handler.CreateTaskRequest:
    properties:
      assignee_id:
        type: string
      description:
        maxLength: 2000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - pending
        - in_progress
        - done
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  internal_handler.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  internal_handler.ListTasksResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_abduromanov2020_tasks-api_internal_usecase.TaskView'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handler.LoginRequest:
    properties:
      email:
        maxLength: 254
        type: string
      password:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      team_name:
        maxLength: 100
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_handler.UpdateTaskRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - pending
        - in_progress
        - done
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - priority
    - status
    - title
    type: object
  internal_handler.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      team_id:
        type: string
    type: object
info:
  contact: {}
  description: Team-scoped task management API with idempotent writes and JWT auth.
  title: Tasks API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Log in and receive a JWT
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.AuthResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Register a new user (creates a team if team_name is empty)
      tags:
      - auth
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Liveness check
      tags:
      - health
  /tasks:
    get:
      parameters:
      - description: filter by status (pending|in_progress|done)
        in: query
        name: status
        type: string
      - description: search by title (substring, case-insensitive)
        in: query
        name: q
        type: string
      - default: 1
        description: 1-based page
        in: query
        name: page
        type: integer
      - default: 10
        description: page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.ListTasksResponse'
      security:
      - ApiKeyAuth: []
      summary: List team tasks with filter, search, pagination
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: UUID v4
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_abduromanov2020_tasks-api_internal_usecase.TaskView'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: IDEMPOTENCY_IN_FLIGHT
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a task (idempotent)
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a task (creator only)
      tags:
      - tasks
    get:
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abduromanov2020_tasks-api_internal_usecase.TaskView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a task by id (team-scoped)
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abduromanov2020_tasks-api_internal_usecase.TaskView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a task (full replace of mutable fields)
      tags:
      - tasks
  /tasks/{id}/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: assignee payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AssignTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abduromanov2020_tasks-api_internal_usecase.TaskView'
        "403":
          description: cross-team assignment refused
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Assign a task to another team member (transactional)
      tags:
      - tasks
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"